+++ app/crud/transaction.py
@@ -3,38 +3,42 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from app.crud import user
 from app.crud import category
 from app.exc.error import (
-    UserSearchError,
     CategorySearchError,
     DateError,
     TransactionCreateError,
+    GetTransactionError,
 )
+from sqlalchemy import Result
 from app.models.transaction import Transaction
 from datetime import date
+from app.models.user import User
 from sqlalchemy import select
 
 
 async def create_transaction_crud(
-    transaction_data: CreateTransaction, session: AsyncSession
+    transaction_data: CreateTransaction,
+    session: AsyncSession,
+    user_db: User,
 ):
-    current_user = await user.get_user_by_id_crud(
-        user_id=transaction_data.user_id, session=session
-    )
-    if current_user is None:
-        raise UserSearchError
+
     current_category = await category.get_category_by_id_crud(
         category_id=transaction_data.category_id, session=session
     )
     if current_category is None:
         raise CategorySearchError
-    if (
-        current_category.user_id != transaction_data.user_id
-        and current_category.user_id is not None
-    ):
+    if current_category.user_id != user_db.id and current_category.user_id is not None:
         raise TransactionCreateError
     current_date = date.today()
     if transaction_data.transaction_date > current_date:
         raise DateError
-    new_transaction = Transaction(**transaction_data.model_dump())
+    new_transaction = Transaction(
+        amount=transaction_data.amount,
+        cost=transaction_data.cost,
+        category_id=transaction_data.category_id,
+        user_id=user_db.id,
+        description=transaction_data.description,
+        transaction_date=transaction_data.transaction_date,
+    )
     session.add(new_transaction)
     await session.commit()
     await session.refresh(new_transaction)
@@ -44,18 +48,34 @@ async def create_transaction_crud(
 async def get_transaction_by_id_crud(
     transaction_id: int,
     session: AsyncSession,
+    user_db: User,
 ):
-    current_transaction = await session.get(Transaction, transaction_id)
-    return current_transaction
+    stmt = (
+        select(Transaction)
+        .filter(Transaction.id == transaction_id)
+        .order_by(Transaction.id)
+    )
+    result: Result = await session.execute(stmt)
+    transations: Transaction = result.scalars().one_or_none()
+    if transations is None:
+        return None
+    if transations.user_id != user_db.id:
+        raise GetTransactionError
+    return transations
 
 
 async def get_list_transactions_crud(
+    user_db: User,
     session: AsyncSession,
     start: int = 0,
     stop: int = 3,
 ):
     stmt = (
-        select(Transaction).order_by(Transaction.id).limit(stop - start).offset(start)
+        select(Transaction)
+        .filter(Transaction.user_id == user_db.id)
+        .order_by(Transaction.id)
+        .limit(stop - start)
+        .offset(start)
     )
     result = await session.execute(stmt)
     transactions: list = result.scalars().all()
@@ -66,9 +86,12 @@ async def update_transaction_crud(
     transaction_data: UpdateTransaction,
     transaction_id: int,
     session: AsyncSession,
+    user_db: User,
 ):
     update_transaction = await get_transaction_by_id_crud(
-        transaction_id=transaction_id, session=session
+        transaction_id=transaction_id,
+        session=session,
+        user_db=user_db,
     )
     if update_transaction is None:
         return None
@@ -78,6 +101,8 @@ async def update_transaction_crud(
         and transaction_data.transaction_date > current_date
     ):
         raise DateError
+    if user_db.id != update_transaction.user_id:
+        raise GetTransactionError
     update_data: dict = transaction_data.model_dump(exclude_unset=True)
     if len(update_data) == 0:
         return update_transaction
@@ -88,12 +113,18 @@ async def update_transaction_crud(
     return update_transaction
 
 
-async def delete_transaction_crud(transaction_id: int, session: AsyncSession):
+async def delete_transaction_crud(
+    transaction_id: int,
+    session: AsyncSession,
+    user_db: User,
+):
     delete_transaction = await get_transaction_by_id_crud(
-        transaction_id=transaction_id, session=session
+        transaction_id=transaction_id, session=session, user_db=user_db
     )
     if delete_transaction is None:
         return None
+    if user_db.id != delete_transaction.user_id:
+        raise GetTransactionError
     await session.delete(delete_transaction)
     await session.commit()
     return {"message": "delete"}
